+++ rocDecode-setup.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2023 - 2024 Advanced Micro Devices, Inc. All rights reserved.
+# Copyright (c) 2023 - 2025 Advanced Micro Devices, Inc. All rights reserved.
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -28,8 +28,8 @@ if sys.version_info[0] < 3:
 else:
     import subprocess
 
-__copyright__ = "Copyright (c) 2023 - 2024, AMD ROCm rocDecode"
-__version__ = "2.4.0"
+__copyright__ = "Copyright (c) 2023 - 2025, AMD ROCm rocDecode"
+__version__ = "2.5.0"
 __email__ = "mivisionx.support@amd.com"
 __status__ = "Shipping"
 
@@ -182,7 +182,6 @@ runtimeDebianPackages = [
     'vainfo'
 ]
 ffmpegDebianPackages = [
-    'ffmpeg',
     'libavcodec-dev',
     'libavformat-dev',
     'libavutil-dev'